音声認識のページをmainからvoiceRecognitionPageへ移動
